+++ include/mruby_integration/models/circle.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "mruby.h"
+
+extern RClass* Circle_class;
+
+void
+append_models_Circle(mrb_state*);