+++ include/mruby_integration/struct_types.hpp
@@ -12,7 +12,8 @@ void
 add_owned_object(void*);
 
 extern mrb_data_type Camera2D_type;
-extern mrb_data_type Colour_type;
+extern mrb_data_type Circle_type;
+extern mrb_data_type Color_type;
 extern mrb_data_type Font_type;
 extern mrb_data_type Image_type;
 extern mrb_data_type Monitor_type;