+++ mrb_doc/models/colour.rb
@@ -1,5 +1,12 @@
 class Colour
+  # @return [Integer]
+  attr_accessor :red, :green, :blue, :alpha
+
   # Creates a new instance of Colour.
+  #
+  # @example Basic usage
+  #   purple = Colour.new(red: 128, green: 0, blue: 255)
+  #
   # @param red [Integer] A value between 0 and 255.
   # @param blue [Integer] A value between 0 and 255.
   # @param green [Integer] A value between 0 and 255.
@@ -9,30 +16,6 @@ class Colour
     # mrb_Colour_initialize
     # src/mruby_integration/models/colour.cpp
     Colour.new
-  end
-
-  def red=(red)
-    # mrb_Colour_set_red
-    # src/mruby_integration/models/colour.cpp
-    255
-  end
-
-  def green=(green)
-    # mrb_Colour_set_green
-    # src/mruby_integration/models/colour.cpp
-    255
-  end
-
-  def blue=(blue)
-    # mrb_Colour_set_blue
-    # src/mruby_integration/models/colour.cpp
-    255
-  end
-
-  def alpha=(alpha)
-    # mrb_Colour_set_alpha
-    # src/mruby_integration/models/colour.cpp
-    255
   end
 
   # Returns a new {Colour} which is a faded version of the original.