+++ mrb_doc/models/shader.rb
@@ -1,4 +1,7 @@
 class Shader
+  # @return [Integer]
+  attr_writer :id
+
   # Creates a new instance of {Shader}.
   # @param id [Integer]
   # @return [Shader]
@@ -6,11 +9,5 @@ class Shader
     # mrb_Shader_initialize
     # src/mruby_integration/models/shader.cpp
     Shader.new
-  end
-
-  def id=(id)
-    # mrb_Shader_set_id
-    # src/mruby_integration/models/shader.cpp
-    0
   end
 end