+++ src/mruby_integration/models.cpp
@@ -1,5 +1,6 @@
 #include "mruby_integration/models/audio.hpp"
 #include "mruby_integration/models/camera2d.hpp"
+#include "mruby_integration/models/circle.hpp"
 #include "mruby_integration/models/colour.hpp"
 #include "mruby_integration/models/cursor.hpp"
 #include "mruby_integration/models/font.hpp"
@@ -22,6 +23,7 @@ append_models(mrb_state* mrb)
 {
   append_models_Audio(mrb);
   append_models_Camera2D(mrb);
+  append_models_Circle(mrb);
   append_models_Colour(mrb);
   append_models_Cursor(mrb);
   append_models_Font(mrb);