+++ src/ruby/models/circle.rb
@@ -1,0 +1,5 @@
+# The {Circle} class is used for drawing circles.
+class Circle
+  # @return [Colour]
+  attr_reader :colour, :outline, :gradient
+end